Soft fix to the build error, added unit tests to buffer.rs
1. Deleted some function calls for the previous TestWriter struct and that seemed to fix the build error.

2. Removed the lifetime parameters from TermWriter

3. Commented out my code in fn main() since it seems unnecessary for now
